requirements.txt
and==66.0.2
asgiref==3.6.0
beautifulsoup4==4.11.2
bootstrap4==0.1.0
dj-database-url==1.2.0
Django==4.1.7
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
python-decouple==3.8
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
